DrinkService 분리
(cherry picked from commit 829bc7cb15d35c1bbbbb121579aae78fb80258d3)
