+++ Mulimee/Mulimee/App/MainView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct MainView: View {
-    @State private var drink: Drink = Drink(drinkRepository: DrinkRepositoryService(),
+    @State private var drink: Drink = Drink(drinkRepository: DrinkRepositoryImpl(),
                                             healthKitRepository: HealthKitRepositoryImpl())
     @State private var isLaunching = true
     private let moveToTimer = Timer.publish(every: 2,on: .main, in: .common).autoconnect().first()
