+++ Mulimee/Mulimee/Drink/Model/Drink.swift
@@ -81,4 +81,3 @@ final class Drink {
         numberOfGlassesSubject.send(glasses - 1)
     }
 }
-
